# Further optimizations