… LuckParser/Models/BossLogic/BossLogic.cs
… LuckParser/Models/BossLogic/Deimos.cs
… LuckParser/Models/BossLogic/Dhuum.cs
… LuckParser/Models/BossLogic/FractalLogic.cs
… LuckParser/Models/BossLogic/Gorseval.cs
… LuckParser/Models/BossLogic/KeepConstruct.cs
… LuckParser/Models/BossLogic/Matthias.cs
… LuckParser/Models/BossLogic/Sabetha.cs
… LuckParser/Models/BossLogic/Samarog.cs
… LuckParser/Models/BossLogic/Slothasor.cs
… LuckParser/Models/BossLogic/SoullessHorror.cs
… LuckParser/Models/BossLogic/ValeGuardian.cs
… LuckParser/Models/BossLogic/Xera.cs
… LuckParser/Models/ParseModels/Boons/Boon.cs
… LuckParser/Models/ParseModels/CombatData.cs
… LuckParser/Models/ParseModels/CombatItem.cs
… LuckParser/Models/ParseModels/Logs/DamageLog.cs
… LuckParser/Models/ParseModels/Logs/DamageLogCondition.cs
… LuckParser/Models/ParseModels/Logs/DamageLogPower.cs
… LuckParser/Models/ParseModels/Players/AbstractMasterPlayer.cs
… LuckParser/Models/ParseModels/Players/AbstractPlayer.cs
… LuckParser/Models/ParseModels/Players/Boss.cs
… LuckParser/Models/ParseModels/Players/Minion.cs
… LuckParser/Models/ParseModels/Players/Player.cs