Includes a change for CombatData that changes getters/setters into properties (as that apparently helps the compiler/JIT to optimize it better as well).

Includes a fix for duplicated boon ids from copy-paste mistakes (needed for a dictionary which requires unique skill ids)

Brings down total log processing time to ~83% of original, most improvements are in statistics generation (78% of original time)

Log	| Before (ms)	| GetStates optimized (ms)	|  CombatItem Properties (ms)	|SetPlayerBoons optimized (ms)
--- | --- | --- | --- | ---
cairn_20180809-052839	|1173.479	|1071.3064	|1018.0766	|949.0993
cairn_20180814-040825	|1139.8919	|1055.2094	|996.7005	|920.136
cairn_20180820-201521	|878.209	|813.2028	|774.5726	|713.4486
deimos_20180814-051155	|4420.4371	|4239.8768	|4105.6815	|3793.6144
deimos_20180821-000855	|4823.1289	|4459.9949	|4207.3441	|3851.7822
deimos_20180821-214256	|4242.0258	|3826.7553	|3658.0624	|3380.5514
dhuum_20180813-190655	|3957.9583	|3667.7938	|3460.0617	|3282.447
dhuum_20180821-024226	|4494.3554	|4239.9547	|3984.0434	|3749.703
dhuumcm_20180813-045358	|3771.4957	|3537.7284	|3338.321	|3146.1439
gorseval_20180807-010122	|3203.7422	|2991.9953	|2837.7855	|2682.4804
gorseval_20180813-105556	|2939.3652	|2625.0048	|2538.4232	|2387.4989
gorseval_20180821-203347	|3383.3457	|3118.5756	|3042.4097	|2832.0022
kc_20180807-050104	|3433.8174	|3257.8355	|3125.3202	|2943.1345
kc_20180814-033936	|4125.0955	|3875.1236	|3682.4597	|3458.5667
kc_20180821-003658	|4504.3317	|4169.4485	|3994.2165	|3789.9068
matt_20180817-054149	|4373.1702	|4101.2666	|4039.7088	|3726.8053
matt_20180821-015030	|5471.9062	|5253.6901	|4999.0323	|4696.7011
matt_20180822-215817	|6285.5241	|6020.1077	|5606.4318	|5331.549
mo_20180816-203538	|934.206	|876.6116	|832.6629	|767.2111
mo_20180819-174250	|1021.0612	|962.332	|918.5514	|860.9997
mo_20180820-202322	|1039.2315	|995.2919	|943.3747	|869.7492
sabetha_20180813-112032	|4974.406	|4705.8742	|4422.0585	|4172.7523
sabetha_20180817-044932	|4963.9244	|4760.4706	|4515.0608	|4329.5339
sabetha_20180821-210426	|5452.9707	|5371.0402	|4951.509	|4750.4955
samarog_20180814-042916	|4360.3985	|4236.6957	|3899.5267	|3716.4354
samarog_20180816-205700	|5761.201	|5523.8626	|5239.051	|4950.4988
samarog_20180820-233933	|3193.089	|2997.3576	|2850.615	|2650.6243
sh_20180806-095250	|2411.5156	|2261.6379	|2093.8042	|1940.8359
sh_20180807-212636	|1956.6099	|1820.1641	|1677.3868	|1560.8656
sh_20180814-054151	|2757.4396	|2643.5993	|2470.6831	|2304.6528
sloth_20180817-050944	|1290.0775	|1229.4549	|1127.0108	|1039.3227
sloth_20180821-012528	|988.3265	|906.6352	|863.4733	|794.9065
sloth_20180822-211108	|1280.4036	|1228.9624	|1109.8327	|1025.0684
vg_20180807-004529	|2788.8968	|2634.5437	|2458.1645	|2350.0541
vg_20180813-103904	|2682.627	|2540.6337	|2350.9403	|2232.2122
vg_20180820-213803	|3105.5177	|2940.3016	|2779.1675	|2614.2465
xera_20180807-051530	|2269.3049	|2084.7061	|1982.5673	|1878.2533
xera_20180814-040024	|3342.9901	|3191.0856	|2962.8863	|2802.4343
xera_20180821-011535	|2125.7065	|2039.7104	|1888.7679	|1787.313
Average | 3213.3636 | 3032.7138 | 2865.2755 | 2693.1804